move DomainCard to /domains
